rename RenderingComponent -> Component
