+++ src/components/notification.js
@@ -1,4 +1,4 @@
-import { RenderingComponent, createElement } from './RenderingComponent';
+import { Component, createElement } from './Component';
 import warningSvg from '../img/warning.min.svg';
 
 const createNotificationItemTemplate = (violation) => `
@@ -20,7 +20,7 @@ const createNotificationListTemplate = (violationInformation) => {
   return `<div class="notification-list">${notificationItemTemplates}</div>`;
 };
 
-export class Notification extends RenderingComponent {
+export class Notification extends Component {
   getElement() {
     if (this.element === null) {
       this.element = createElement(createNotificationListTemplate(this.data));
